// client/stocklinx app/src/redux/fieldSetCustomField/constant.ts
export enum fieldSetCustomFieldConst {
  FETCH_FIELDSETCUSTOMFIELDS_REQUEST = "FETCH_FIELDSETCUSTOMFIELDS_REQUEST",
  FETCH_FIELDSETCUSTOMFIELDS_SUCCESS = "FETCH_FIELDSETCUSTOMFIELDS_SUCCESS",
  FETCH_FIELDSETCUSTOMFIELDS_FAILURE = "FETCH_FIELDSETCUSTOMFIELDS_FAILURE",
  FETCH_FIELDSETCUSTOMFIELD_REQUEST = "FETCH_FIELDSETCUSTOMFIELD_REQUEST",
  FETCH_FIELDSETCUSTOMFIELD_SUCCESS = "FETCH_FIELDSETCUSTOMFIELD_SUCCESS",
  FETCH_FIELDSETCUSTOMFIELD_FAILURE = "FETCH_FIELDSETCUSTOMFIELD_FAILURE",
  FETCH_FIELDSETCUSTOMFIELD_COUNTS_REQUEST = "FETCH_FIELDSETCUSTOMFIELD_COUNTS_REQUEST",
  FETCH_FIELDSETCUSTOMFIELD_COUNTS_SUCCESS = "FETCH_FIELDSETCUSTOMFIELD_COUNTS_SUCCESS",
  FETCH_FIELDSETCUSTOMFIELD_COUNTS_FAILURE = "FETCH_FIELDSETCUSTOMFIELD_COUNTS_FAILURE",
  CREATE_FIELDSETCUSTOMFIELD_REQUEST = "CREATE_FIELDSETCUSTOMFIELD_REQUEST",
  CREATE_FIELDSETCUSTOMFIELD_SUCCESS = "CREATE_FIELDSETCUSTOMFIELD_SUCCESS",
  CREATE_FIELDSETCUSTOMFIELD_FAILURE = "CREATE_FIELDSETCUSTOMFIELD_FAILURE",
  CREATE_RANGE_FIELDSETCUSTOMFIELD_REQUEST = "CREATE_RANGE_FIELDSETCUSTOMFIELD_REQUEST",
  CREATE_RANGE_FIELDSETCUSTOMFIELD_SUCCESS = "CREATE_RANGE_FIELDSETCUSTOMFIELD_SUCCESS",
  CREATE_RANGE_FIELDSETCUSTOMFIELD_FAILURE = "CREATE_RANGE_FIELDSETCUSTOMFIELD_FAILURE",
  UPDATE_FIELDSETCUSTOMFIELD_REQUEST = "UPDATE_FIELDSETCUSTOMFIELD_REQUEST",
  UPDATE_FIELDSETCUSTOMFIELD_SUCCESS = "UPDATE_FIELDSETCUSTOMFIELD_SUCCESS",
  UPDATE_FIELDSETCUSTOMFIELD_FAILURE = "UPDATE_FIELDSETCUSTOMFIELD_FAILURE",
  REMOVE_FIELDSETCUSTOMFIELD_REQUEST = "REMOVE_FIELDSETCUSTOMFIELD_REQUEST",
  REMOVE_FIELDSETCUSTOMFIELD_SUCCESS = "REMOVE_FIELDSETCUSTOMFIELD_SUCCESS",
  REMOVE_FIELDSETCUSTOMFIELD_FAILURE = "REMOVE_FIELDSETCUSTOMFIELD_FAILURE",
  REMOVE_RANGE_FIELDSETCUSTOMFIELD_REQUEST = "REMOVE_RANGE_FIELDSETCUSTOMFIELD_REQUEST",
  REMOVE_RANGE_FIELDSETCUSTOMFIELD_SUCCESS = "REMOVE_RANGE_FIELDSETCUSTOMFIELD_SUCCESS",
  REMOVE_RANGE_FIELDSETCUSTOMFIELD_FAILURE = "REMOVE_RANGE_FIELDSETCUSTOMFIELD_FAILURE",
  SYNCHRONIZE_FIELDSETCUSTOMFIELD_REQUEST = "SYNCHRONIZE_FIELDSETCUSTOMFIELD_REQUEST",
  SYNCHRONIZE_FIELDSETCUSTOMFIELD_SUCCESS = "SYNCHRONIZE_FIELDSETCUSTOMFIELD_SUCCESS",
  SYNCHRONIZE_FIELDSETCUSTOMFIELD_FAILURE = "SYNCHRONIZE_FIELDSETCUSTOMFIELD_FAILURE",
  SET_FIELDSETCUSTOMFIELD = "SET_FIELDSETCUSTOMFIELD",
  CLEAR_FIELDSETCUSTOMFIELD = "CLEAR_FIELDSETCUSTOMFIELD",
  SET_FIELDSETCUSTOMFIELDS = "SET_FIELDSETCUSTOMFIELDS",
  CLEAR_FIELDSETCUSTOMFIELDS = "CLEAR_FIELDSETCUSTOMFIELDS",
}